validation tests for investments associated with client
